T285 Revert back to single file for routes
